Extract GUI menu/text helpers and fixture update type mapping
