+++ gui/fixturetablepanel.cpp
@@ -16,6 +16,7 @@
  * along with Perastage. If not, see <https://www.gnu.org/licenses/>.
  */
 #include "fixturetablepanel.h"
+#include "fixturetablepanel_update_types.h"
 #include "addressdialog.h"
 #include "configmanager.h"
 #include "selection_origin_token.h"
@@ -139,57 +140,6 @@ bool ShouldRefreshFixtureSummary(
     return false;
   }
   return true;
-}
-
-FixtureTablePanel::SceneDataUpdateType UpdateTypeForColumnImpl(int column) {
-  using SceneDataUpdateType = FixtureTablePanel::SceneDataUpdateType;
-  switch (column) {
-  case 1:
-    return SceneDataUpdateType::kVisualLabelOnly;
-  case 5:
-  case 6:
-  case 8:
-    return SceneDataUpdateType::kPatchOnly;
-  case 10:
-  case 11:
-  case 12:
-  case 13:
-  case 14:
-  case 15:
-    return SceneDataUpdateType::kTransformOnly;
-  case 16:
-  case 17:
-    return SceneDataUpdateType::kWeightOrPosition;
-  case 18:
-    return SceneDataUpdateType::kCategoryOnly;
-  case 19:
-    return SceneDataUpdateType::kAppearanceOnly;
-  case 0:
-    return SceneDataUpdateType::kFixtureIdOnly;
-  case 2:
-  case 3:
-  case 4:
-  case 7:
-  case 9:
-    return SceneDataUpdateType::kMetadataOnly;
-  default:
-    return SceneDataUpdateType::kGeneral;
-  }
-}
-
-FixtureTablePanel::SceneDataUpdateType CombineUpdateTypesImpl(
-    FixtureTablePanel::SceneDataUpdateType lhs,
-    FixtureTablePanel::SceneDataUpdateType rhs) {
-  using SceneDataUpdateType = FixtureTablePanel::SceneDataUpdateType;
-  if (lhs == rhs)
-    return lhs;
-  if (lhs == SceneDataUpdateType::kGeneral || rhs == SceneDataUpdateType::kGeneral)
-    return SceneDataUpdateType::kGeneral;
-  if (lhs == SceneDataUpdateType::kVisualLabelOnly)
-    return rhs;
-  if (rhs == SceneDataUpdateType::kVisualLabelOnly)
-    return lhs;
-  return SceneDataUpdateType::kGeneral;
 }
 
 bool IsRedCell(const ColorfulDataViewListStore *store, int row, int col) {
@@ -1153,11 +1103,13 @@ void FixtureTablePanel::SetInstance(FixtureTablePanel *panel) {
   s_instance = panel;
 }
 
+// Maps a table column index to its corresponding scene update category.
 FixtureTablePanel::SceneDataUpdateType
 FixtureTablePanel::UpdateTypeForColumn(int column) {
   return UpdateTypeForColumnImpl(column);
 }
 
+// Combines two scene update categories into a safe aggregate update.
 FixtureTablePanel::SceneDataUpdateType FixtureTablePanel::CombineUpdateTypes(
     SceneDataUpdateType lhs, SceneDataUpdateType rhs) {
   return CombineUpdateTypesImpl(lhs, rhs);
