+++ gui/mainwindow_menu.cpp
@@ -19,6 +19,7 @@
 #include "mainwindow/controllers/mainwindow_io_controller.h"
 #include "mainwindow_view_controller.h"
 #include "mainwindow_menu_builders.h"
+#include "mainwindow_menu_text_utils.h"
 
 #include <algorithm>
 #include <chrono>
@@ -90,27 +91,11 @@
 
 namespace {
 
-// Converts a wxString value to a UTF-8 std::string.
-std::string WxToUtf8(const wxString &value) {
-  const wxScopedCharBuffer utf8 = value.ToUTF8();
-  if (utf8)
-    return std::string(utf8.data(), utf8.length());
-  return value.ToStdString();
-}
-
 struct HelpMarkdown {
   std::string english;
   std::string spanish;
   bool hasSections = false;
 };
-
-// Removes leading blank-line whitespace from a text block.
-std::string TrimLeadingWhitespace(const std::string &text) {
-  const auto start = text.find_first_not_of("\r\n");
-  if (start == std::string::npos)
-    return std::string();
-  return text.substr(start);
-}
 
 // Splits help markdown into English and Spanish sections when markers are present.
 HelpMarkdown SplitHelpMarkdown(const std::string &markdown) {
@@ -159,11 +144,6 @@ HelpMarkdown SplitHelpMarkdown(const std::string &markdown) {
   return result;
 }
 
-// Wraps HTML body content in a UTF-8 HTML document shell.
-std::string WrapHelpHtml(const std::string &body) {
-  return "<html><head><meta charset=\"UTF-8\"></head><body>" + body +
-         "</body></html>";
-}
 } // namespace
 
 // Builds and registers the main application toolbars.
